Moved custom exceptions into a seperate module
