+++ virtual_rainforest/main.py
@@ -10,13 +10,10 @@ from typing import Any, Type, Union
 import pint
 from numpy import datetime64, timedelta64
 
-from virtual_rainforest.core.base_model import (
-    MODEL_REGISTRY,
-    BaseModel,
-    InitialisationError,
-)
+from virtual_rainforest.core.base_model import MODEL_REGISTRY, BaseModel
 from virtual_rainforest.core.config import validate_config
 from virtual_rainforest.core.data import Data
+from virtual_rainforest.core.exceptions import InitialisationError
 from virtual_rainforest.core.grid import Grid
 from virtual_rainforest.core.logger import LOGGER
